Added timeout to prevent roslisp builds from blocking forever
Sometimes, marker files that indicate compilation of a specific file don't get
removed for some strange reason. This leads to infinite blocking of lisp
compilation and the only fix was to remove the marker file by hand. Added a
timeout that, when it expires, removes the marker file automatically and the
build can continue.