Ensure that the compilation-marker-file-path introduces no new directories
If an ASDF component has a complex name (includes a /) character, then the
compilation marker file will be placed in a subdirectory of ROS home. This
results in `open` raising a `file-error` if the directory doesn't exist (it
will never exist without external intervention), causing the file to never be
compiled.

This fixes the issue by replacing all / characters with _ characters in the
ASDF component name.